models/model_manager.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
模型管理模块
"""

import pandas as pd
import numpy as np
from typing import Dict, List, Optional, Union, Any
from abc import ABC, abstractmethod
import joblib
import pickle
from sklearn.model_selection import TimeSeriesSplit, cross_val_score
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score

from .ml_models import LightGBMModel, XGBoostModel, LinearModel, NeuralNetworkModel
from .ensemble_models import StackingEnsemble, BlendingEnsemble, VotingEnsemble
from .time_series_models import ARIMAModel, LSTMModel, GRUModel


class BaseModel(ABC):
    """模型基类"""
    
    def __init__(self, name: str, params: Dict = None):
        """初始化模型"""
        pass
    
    @abstractmethod
    def fit(self, X: pd.DataFrame, y: pd.Series, sample_weight: pd.Series = None):
        """训练模型"""
        pass
    
    @abstractmethod
    def predict(self, X: pd.DataFrame) -> np.ndarray:
        """预测"""
        pass
    
    @abstractmethod
    def predict_proba(self, X: pd.DataFrame) -> np.ndarray:
        """预测概率"""
        pass
    
    def get_feature_importance(self) -> pd.Series:
        """获取特征重要性"""
        pass
    
    def save_model(self, file_path: str):
        """保存模型"""
        pass
    
    def load_model(self, file_path: str):
        """加载模型"""
        pass


class ModelEvaluator:
    """模型评估器"""
    
    def __init__(self):
        """初始化评估器"""
        pass
    
    def evaluate_regression(self, y_true: np.ndarray, y_pred: np.ndarray) -> Dict[str, float]:
        """回归模型评估"""
        pass
    
    def evaluate_classification(self, y_true: np.ndarray, y_pred: np.ndarray) -> Dict[str, float]:
        """分类模型评估"""
        pass
    
    def calculate_ic(self, y_true: np.ndarray, y_pred: np.ndarray) -> float:
        """计算信息系数"""
        pass
    
    def calculate_rank_ic(self, y_true: np.ndarray, y_pred: np.ndarray) -> float:
        """计算排序信息系数"""
        pass
    
    def calculate_sharpe_ratio(self, returns: np.ndarray) -> float:
        """计算夏普比率"""
        pass
    
    def calculate_max_drawdown(self, cumulative_returns: np.ndarray) -> float:
        """计算最大回撤"""
        pass
    
    def cross_validate(self, model: BaseModel, X: pd.DataFrame, y: pd.Series,
                      cv_method: str = "time_series", n_splits: int = 5) -> Dict[str, float]:
        """交叉验证"""
        pass


class HyperparameterOptimizer:
    """超参数优化器"""
    
    def __init__(self, method: str = "bayesian"):
        """初始化优化器"""
        pass
    
    def optimize(self, model_class, X: pd.DataFrame, y: pd.Series,
                param_space: Dict, n_trials: int = 100) -> Dict:
        """超参数优化"""
        pass
    
    def grid_search(self, model_class, X: pd.DataFrame, y: pd.Series,
                   param_grid: Dict) -> Dict:
        """网格搜索"""
        pass
    
    def random_search(self, model_class, X: pd.DataFrame, y: pd.Series,
                     param_distributions: Dict, n_iter: int = 100) -> Dict:
        """随机搜索"""
        pass
    
    def bayesian_optimization(self, model_class, X: pd.DataFrame, y: pd.Series,
                             param_space: Dict, n_trials: int = 100) -> Dict:
        """贝叶斯优化"""
        pass


class FeatureSelector:
    """特征选择器"""
    
    def __init__(self):
        """初始化特征选择器"""
        pass
    
    def select_by_importance(self, model: BaseModel, X: pd.DataFrame,
                           top_k: int = 50) -> List[str]:
        """基于重要性选择特征"""
        pass
    
    def select_by_correlation(self, X: pd.DataFrame, y: pd.Series,
                            threshold: float = 0.05) -> List[str]:
        """基于相关性选择特征"""
        pass
    
    def select_by_mutual_info(self, X: pd.DataFrame, y: pd.Series,
                             top_k: int = 50) -> List[str]:
        """基于互信息选择特征"""
        pass
    
    def recursive_feature_elimination(self, model: BaseModel, X: pd.DataFrame,
                                    y: pd.Series, n_features: int = 50) -> List[str]:
        """递归特征消除"""
        pass
    
    def forward_selection(self, model: BaseModel, X: pd.DataFrame,
                         y: pd.Series, max_features: int = 50) -> List[str]:
        """前向选择"""
        pass


class ModelTrainer:
    """模型训练器"""
    
    def __init__(self):
        """初始化训练器"""
        pass
    
    def train_single_model(self, model: BaseModel, X: pd.DataFrame,
                          y: pd.Series, validation_data: tuple = None) -> BaseModel:
        """训练单个模型"""
        pass
    
    def train_ensemble_model(self, base_models: List[BaseModel], X: pd.DataFrame,
                           y: pd.Series, ensemble_method: str = "stacking") -> BaseModel:
        """训练集成模型"""
        pass
    
    def online_learning(self, model: BaseModel, new_X: pd.DataFrame,
                       new_y: pd.Series) -> BaseModel:
        """在线学习"""
        pass
    
    def incremental_training(self, model: BaseModel, new_data: pd.DataFrame,
                           retrain_threshold: float = 0.1) -> BaseModel:
        """增量训练"""
        pass


class ModelManager:
    """模型管理主类"""
    
    def __init__(self, config: Dict = None):
        """初始化模型管理器"""
        pass
    
    def initialize(self, config: Dict):
        """初始化配置"""
        pass
    
    def create_model(self, model_type: str, params: Dict = None) -> BaseModel:
        """创建模型"""
        pass
    
    def train_models(self, X: pd.DataFrame, y: pd.DataFrame,
                    model_configs: List[Dict]) -> Dict[str, BaseModel]:
        """训练多个模型"""
        pass
    
    def select_best_model(self, models: Dict[str, BaseModel],
                         X_val: pd.DataFrame, y_val: pd.Series,
                         metric: str = "ic") -> str:
        """选择最佳模型"""
        pass
    
    def ensemble_models(self, models: Dict[str, BaseModel],
                       method: str = "weighted_average",
                       weights: Dict[str, float] = None) -> BaseModel:
        """模型集成"""
        pass
    
    def predict_multi_horizon(self, models: Dict[str, BaseModel],
                             X: pd.DataFrame, horizons: List[int]) -> pd.DataFrame:
        """多周期预测"""
        pass
    
    def retrain_models(self, models: Dict[str, BaseModel],
                      new_data: pd.DataFrame, retrain_schedule: Dict):
        """重训练模型"""
        pass
    
    def validate_models(self, models: Dict[str, BaseModel],
                       validation_data: pd.DataFrame) -> Dict[str, Dict]:
        """验证模型"""
        pass
    
    def get_model_performance(self, model: BaseModel, X: pd.DataFrame,
                             y: pd.Series) -> Dict[str, float]:
        """获取模型性能"""
        pass
    
    def save_models(self, models: Dict[str, BaseModel], save_dir: str):
        """保存模型"""
        pass
    
    def load_models(self, model_dir: str) -> Dict[str, BaseModel]:
        """加载模型"""
        pass
    
    def monitor_model_drift(self, model: BaseModel, X_train: pd.DataFrame,
                           X_current: pd.DataFrame) -> Dict[str, float]:
        """监控模型漂移"""
        pass
    
    def auto_model_selection(self, X: pd.DataFrame, y: pd.Series,
                           model_types: List[str] = None) -> BaseModel:
        """自动模型选择"""
        pass
    
    def explain_model(self, model: BaseModel, X: pd.DataFrame,
                     method: str = "shap") -> Dict:
        """模型解释"""
        pass 