requirements.txt
# 数据处理和分析
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0

# 机器学习
scikit-learn>=1.1.0
lightgbm>=3.3.0
xgboost>=1.6.0
catboost>=1.0.0
tensorflow>=2.8.0
torch>=1.12.0

# 金融数据源
akshare>=1.8.0
tushare>=1.2.89
yfinance>=0.1.87

# 技术分析
talib>=0.4.25
ta>=0.10.0

# 优化算法
cvxpy>=1.2.0
cvxopt>=1.3.0
pypfopt>=1.5.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.8.0
bokeh>=2.4.0

# 统计分析
statsmodels>=0.13.0
arch>=5.3.0
pymc>=4.0.0

# 数据存储
h5py>=3.7.0
tables>=3.7.0
pymongo>=4.1.0
redis>=4.3.0
sqlalchemy>=1.4.0

# 并行计算
joblib>=1.1.0
dask>=2022.5.0
ray>=1.13.0

# 配置管理
pyyaml>=6.0
configparser>=5.2.0
python-dotenv>=0.20.0

# 日志和监控
loguru>=0.6.0
prometheus-client>=0.14.0
grafana-api>=1.0.3

# 网络请求
requests>=2.28.0
aiohttp>=3.8.0

# 时间处理
pytz>=2022.1
arrow>=1.2.0

# 数学计算
sympy>=1.10.0
numba>=0.56.0

# 报告生成
jinja2>=3.1.0
reportlab>=3.6.0
fpdf2>=2.5.0

# 邮件和通知
smtplib
email-validator>=1.2.0

# 测试
pytest>=7.1.0
pytest-cov>=3.0.0

# 开发工具
jupyter>=1.0.0
notebook>=6.4.0
jupyterlab>=3.4.0

# 类型检查
mypy>=0.971
typing-extensions>=4.2.0 